# Finished sample/Generated code/aspnetcoreapi/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Library API",
    "description" : "Through this API you can access authors and their books.",
    "contact" : {
      "name" : "Kevin Dockx",
      "url" : "https://www.twitter.com/KevinDockx",
      "email" : "kevin.dockx@gmail.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/api/v1/authors" : {
      "get" : {
        "tags" : [ "Authors" ],
        "summary" : "Get a list of authors",
        "operationId" : "GetAuthors",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "uniqueItems" : false,
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Author"
                  }
                }
              },
              "application/xml" : {
                "schema" : {
                  "uniqueItems" : false,
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Author"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500" : {
            "description" : "Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/authors/{authorId}" : {
      "get" : {
        "tags" : [ "Authors" ],
        "summary" : "Get an author by his/her id",
        "operationId" : "GetAuthor",
        "parameters" : [ {
          "name" : "authorId",
          "in" : "path",
          "description" : "The id of the author you want to get",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Author"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Author"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500" : {
            "description" : "Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Authors" ],
        "summary" : "Update an author",
        "operationId" : "UpdateAuthor",
        "parameters" : [ {
          "name" : "authorId",
          "in" : "path",
          "description" : "The id of the author to update",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AuthorForUpdate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Author"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Author"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422" : {
            "description" : "Validation error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "500" : {
            "description" : "Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "Authors" ],
        "summary" : "Partially update an author",
        "description" : "Sample request (this request updates the author's **first name**)  \r\n            \r\n            PATCH /authors/authorId\r\n            [ \r\n                {\r\n                    \"op\": \"replace\", \r\n                    \"path\": \"/firstname\", \r\n                    \"value\": \"new first name\" \r\n                } \r\n            ]",
        "operationId" : "UpdateAuthor",
        "parameters" : [ {
          "name" : "authorId",
          "in" : "path",
          "description" : "The id of the author you want to get",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json-patch+json" : {
              "schema" : {
                "uniqueItems" : false,
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Operation"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Returns the updated author",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Author"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Author"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422" : {
            "description" : "Client Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "500" : {
            "description" : "Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProblemDetails" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string"
          }
        },
        "additionalProperties" : {
          "type" : "object"
        }
      },
      "Void" : {
        "type" : "object"
      },
      "Author" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "The id of the author",
            "format" : "uuid"
          },
          "firstName" : {
            "type" : "string",
            "description" : "The first name of the  \r\n**author**"
          },
          "lastName" : {
            "type" : "string",
            "description" : "The last name of the author"
          }
        },
        "description" : "An author with Id, FirstName and LastName fields"
      },
      "AuthorForUpdate" : {
        "required" : [ "firstName", "lastName" ],
        "type" : "object",
        "properties" : {
          "firstName" : {
            "maxLength" : 150,
            "type" : "string",
            "description" : "The first name of the author"
          },
          "lastName" : {
            "maxLength" : 150,
            "type" : "string",
            "description" : "The last name of the author"
          }
        },
        "description" : "An author for update with FirstName and LastName fields"
      },
      "ValidationProblemDetails" : {
        "type" : "object",
        "properties" : {
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "uniqueItems" : false,
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            },
            "readOnly" : true
          },
          "type" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string"
          }
        },
        "additionalProperties" : {
          "type" : "object"
        }
      },
      "Operation" : {
        "type" : "object",
        "properties" : {
          "value" : {
            "type" : "object"
          },
          "path" : {
            "type" : "string"
          },
          "op" : {
            "type" : "string"
          },
          "from" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}